pyproject.toml
[tool.poetry]
name = "k8s-app-abstraction"
version = "0.1.0"
description = ""
authors = ["Darwin Monroy <contact@darwinmonroy.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"
pydantic = "^1.9.0"
kubernetes = "^23.3.0"
parameterize = "^0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.10.1"
black = "^22.1.0"
pre-commit = "^2.17.0"
pytest-cov = "^3.0.0"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target_version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs          # exclude a few common directories in the
    | \.git           # root of the project
    | \.pytest_cache
    | \.mypy_cache
    | \.venv
    | venv
    | _build
    | build
    | dist
    | pynguin.egg-info
  )/
)
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov"
testpaths = [
    "tests",
]

[tool.coverage.run]
branch = true
source = ["k8s_app_abstraction", "tests"]


[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "def __str__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "if typing.TYPE_CHECKING:"
]
